+++ module/actes-generique/action/SAEEnvoiActes.class.php
@@ -42,8 +42,6 @@ class SAEEnvoiActes extends ActionExecutor
         $fluxData = new FluxDataSedaActes($donneesFormulaire);
         $bordereau = $actesSEDA->getBordereau($fluxData);
         $donneesFormulaire->addFileFromData('sae_bordereau', "bordereau.xml", $bordereau);
-        $transferId = $sae->getTransferId($bordereau);
-        $donneesFormulaire->setData("sae_transfert_id", $transferId);
 
         try {
             $actesSEDA->validateBordereau($bordereau);
@@ -61,13 +59,14 @@ class SAEEnvoiActes extends ActionExecutor
 
         $donneesFormulaire->addFileFromCopy('sae_archive', "archive.tar.gz", $archive_path);
 
-        $result = $sae->sendArchive($bordereau, $archive_path);
+        $result = $sae->sendSIP($bordereau, $archive_path);
 
         if (!$result) {
             $this->setLastMessage("L'envoi du bordereau a échoué : " . $sae->getLastError());
             return false;
         }
 
+        $donneesFormulaire->setData('sae_transfert_id', $result);
         $this->addActionOK("Le document a été envoyé au SAE");
         $this->notify($this->action, $this->type, "Le document a été envoyé au SAE");